requirements.txt
asgiref==3.8.1
astroid==2.15.8
black==24.4.2
click==8.1.7
Django==5.0.6
django-cors-headers==4.4.0
django-jazzmin==3.0.0
django-treenode==0.22.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
flake8==6.1.0
flake8-django==1.4
gunicorn==22.0.0
isort==5.13.2
lazy-object-proxy==1.10.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.1.0
PyJWT==2.8.0
python-dotenv==1.0.1
sqlparse==0.5.0
wrapt==1.16.0
